对话

+++ Classes/NpcRelationship.h
@@ -17,23 +17,23 @@ public:
 	void remove_npc(const string& npcName);
 
 	// 设置npc之间的关系
-	void setRelationship(const string& npcA, const string& npcB, int value);
+	void setRelationship(const string& npcA, const string& npcB, double value);
 
 	// 增加/减少关系  
-	void increaseRelationship(const std::string& npcA, const std::string& npcB, int amount);
-	void decreaseRelationship(const std::string& npcA, const std::string& npcB, int amount);
+	void increaseRelationship(const std::string& npcA, const std::string& npcB, double amount);
+	void decreaseRelationship(const std::string& npcA, const std::string& npcB, double amount);
 
 	// 获取关系层级  
 	std::string getRelationshipLevel(const std::string& npcA, const std::string& npcB) const;
 
 	// 获取npc之间的关系
-	int getRelationship(const string& npcA, const string& npcB) const;
+	double getRelationship(const string& npcA, const string& npcB) const;
 
 	// 打印所有npc之间的关系(用于调试)
 	void printRelationships() const;
 
 private:
-	unordered_map<string, unordered_map<string, int>>relationships;
+	unordered_map<string, unordered_map<string, double>>relationships;
 
 	// 关系值上下限  
 	static const int MIN_RELATIONSHIP = -100;
